Colocate Risk service API response types
